Changed as per the PR review
